LayoutTests/platform/mac-mountainlion/platform/mac/fast/loader/file-url-mimetypes-expected.txt

example-3fr.3fr has MIME type application/octet-stream
example-3g2.3g2 has MIME type video/3gpp2
example-3gp.3gp has MIME type video/3gpp
example-3gp2.3gp2 has MIME type video/3gpp2
example-3gpp.3gpp has MIME type video/3gpp
example-Z.Z has MIME type application/x-compress
example-aaf.aaf has MIME type application/octet-stream
example-ac3.ac3 has MIME type audio/ac3
example-ada.ada has MIME type text/plain
example-adb.adb has MIME type text/plain
example-ads.ads has MIME type text/plain
example-ai.ai has MIME type application/postscript
example-aif.aif has MIME type audio/aiff
example-aifc.aifc has MIME type audio/aiff
example-aiff.aiff has MIME type audio/aiff
example-app.app has MIME type application/octet-stream
example-applescript.applescript has MIME type text/plain
example-arw.arw has MIME type application/octet-stream
example-as.as has MIME type application/octet-stream
example-asc.asc has MIME type text/plain
example-asf.asf has MIME type video/x-ms-asf
example-asx.asx has MIME type video/x-ms-asx
example-au.au has MIME type audio/basic
example-aupreset.aupreset has MIME type application/xml
example-avi.avi has MIME type video/avi
example-bcpio.bcpio has MIME type application/x-bcpio
example-bin.bin has MIME type application/macbinary
example-bmp.bmp has MIME type image/bmp
example-bz.bz has MIME type application/x-bzip2
example-bz2.bz2 has MIME type application/x-bzip2
example-c.c has MIME type text/plain
example-c++.c++ has MIME type text/plain
example-caf.caf has MIME type application/octet-stream
example-cc.cc has MIME type text/plain
example-cdf.cdf has MIME type application/x-netcdf
example-class.class has MIME type application/octet-stream
example-cp.cp has MIME type text/plain
example-cpgz.cpgz has MIME type application/x-gzip
example-cpio.cpio has MIME type application/x-cpio
example-cpp.cpp has MIME type text/plain
example-cr2.cr2 has MIME type application/octet-stream
example-crw.crw has MIME type image/x-canon-crw
example-csh.csh has MIME type application/x-csh
example-css.css has MIME type text/css
example-cxx.cxx has MIME type text/plain
example-dcr.dcr has MIME type application/x-director
example-defs.defs has MIME type text/plain
example-dfont.dfont has MIME type application/octet-stream
example-dif.dif has MIME type video/x-dv
example-dir.dir has MIME type application/x-director
example-dll.dll has MIME type application/x-msdownload
example-dls.dls has MIME type audio/dls
example-dlyan.dlyan has MIME type text/plain
example-dmg.dmg has MIME type application/x-diskcopy
example-dms.dms has MIME type application/octet-stream
example-dng.dng has MIME type application/octet-stream
example-doc.doc has MIME type application/msword
example-docm.docm has MIME type application/vnd.ms-word.document.macroenabled.12
example-docx.docx has MIME type application/vnd.openxmlformats-officedocument.wordprocessingml.document
example-dotm.dotm has MIME type application/vnd.ms-word.template.macroenabled.12
example-dotx.dotx has MIME type application/vnd.openxmlformats-officedocument.wordprocessingml.template
example-dv.dv has MIME type video/x-dv
example-dvi.dvi has MIME type application/x-dvi
example-dxr.dxr has MIME type application/x-director
example-dylib.dylib has MIME type application/octet-stream
example-efx.efx has MIME type image/efax
example-eps.eps has MIME type application/postscript
example-etx.etx has MIME type text/x-setext
example-exe.exe has MIME type application/x-msdownload
example-exp.exp has MIME type text/plain
example-exr.exr has MIME type application/octet-stream
example-ez.ez has MIME type application/andrew-inset
example-f.f has MIME type text/plain
example-f77.f77 has MIME type text/plain
example-f90.f90 has MIME type text/plain
example-f95.f95 has MIME type text/plain
example-fdf.fdf has MIME type application/vnd.fdf
example-fff.fff has MIME type application/octet-stream
example-fla.fla has MIME type application/octet-stream
example-for.for has MIME type text/plain
example-fp.fp has MIME type application/x-filemaker
example-fp2.fp2 has MIME type application/x-filemaker
example-fp3.fp3 has MIME type application/x-filemaker
example-fp4.fp4 has MIME type application/x-filemaker
example-fp5.fp5 has MIME type application/x-filemaker
example-fp6.fp6 has MIME type application/x-filemaker
example-fpx.fpx has MIME type image/fpx
example-gif.gif has MIME type image/gif
example-gtar.gtar has MIME type application/x-gtar
example-gz.gz has MIME type application/x-gzip
example-gzip.gzip has MIME type application/x-gzip
example-h.h has MIME type text/plain
example-h++.h++ has MIME type text/plain
example-hdf.hdf has MIME type application/x-hdf
example-hh.hh has MIME type text/plain
example-hp.hp has MIME type text/plain
example-hpp.hpp has MIME type text/plain
example-hqx.hqx has MIME type application/mac-binhex40
example-htm.htm has MIME type text/html
example-html.html has MIME type text/html
example-hxx.hxx has MIME type text/plain
example-icc.icc has MIME type application/octet-stream
example-ice.ice has MIME type x-conference/x-cooltalk
example-icm.icm has MIME type application/octet-stream
example-icns.icns has MIME type application/octet-stream
example-ico.ico has MIME type image/vnd.microsoft.icon
example-ics.ics has MIME type text/calendar
example-ief.ief has MIME type image/ief
example-iges.iges has MIME type model/iges
example-igs.igs has MIME type model/iges
example-iso.iso has MIME type application/octet-stream
example-j2k.j2k has MIME type image/jp2
example-jar.jar has MIME type application/java-archive
example-jav.jav has MIME type text/plain
example-java.java has MIME type text/plain
example-javascript.javascript has MIME type text/javascript
example-jfx.jfx has MIME type application/octet-stream
example-jhtml.jhtml has MIME type text/html
example-jnlp.jnlp has MIME type application/x-java-jnlp-file
example-jp2.jp2 has MIME type image/jp2
example-jpe.jpe has MIME type image/jpeg
example-jpeg.jpeg has MIME type image/jpeg
example-jpf.jpf has MIME type image/jp2
example-jpg.jpg has MIME type image/jpeg
example-jpx.jpx has MIME type image/jp2
Using an image element to load resources, this test makes sure that every MIME type the loader should know about based solely on file extension in file:// urls is correct.
